user_spec2削除
